backend/auth_service/authentication/urls.py
from django.urls import path
from .views import RegisterView, LoginView, DeleteUserView
from rest_framework_simplejwt.views import (
	TokenObtainPairView,
	TokenRefreshView,
)

urlpatterns = [
	path('register/', RegisterView.as_view(), name='register'),
	path('login/', LoginView.as_view(), name='login'),
	path('delete/', DeleteUserView.as_view(), name='delete'),
	path('token/pair/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
	path('token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
]
